remove app.js module and move into main.js
